veridoc/requirements.txt
Django==2.1.7
django-excel==0.0.10
django-filter==2.1.0
django-mssql==1.8
django-pyodbc-azure==2.1.0.0
django-tastypie==0.14.2
djangorestframework==3.9.1
drf-tracking==1.5.0
drfapikey==0.0.3
lml==0.0.9
Markdown==3.0.1
pyexcel==0.5.10
pyexcel-io==0.5.13
pyexcel-webio==0.1.4
pyodbc==4.0.25
python-dateutil==2.8.0
python-mimeparse==1.6.0
pytz==2018.9
six==1.12.0
texttable==1.6.1
